Reformatted Start method
I moved my code to it's own addFood method!

I don't remember who is doing the Nutritional Info method but I added a method called "analyzeMeal" that can house your code when it's complete. Feel free to rename the method (and method call in the start method) to whatever you want!

I'm working on embedding all our stages into one stage. it may require some recombobulating in the form of panes? But I'm still working on wrapping my mind around how the javaFX heierarchy really works.
