+++ bin/.gitignore
@@ -1,0 +1,3 @@
+/Main.class
+/application.css
+/application/
